requirements.txt
asn1crypto==0.23.0
autopep8==1.2
bcrypt==3.1.4
cached-property==1.4.3
certifi==2018.4.16
cffi==1.11.2
chardet==3.0.4
coverage==4.0
cryptography==2.1.3
decorator==4.0.2
Django==2.0
django-bootstrap3==11.0.0
django-debug-toolbar==1.3.2
django-easy-pjax==1.2.0
django-extra-fields==1.0.0
django-filter==1.1.0
django-lazy-tags==0.3.1
django-registration-redux==2.0
djangorestframework==3.9.2
docker==3.4.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
easy-thumbnails==2.4.1
Fabric3==1.13.1.post1
fabric3-virtualenv==0.3.1
gunicorn==19.3.0
idna==2.6
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
jsonschema==2.6.0
paramiko==2.4.0
path.py==8.1
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
psycopg2==2.8.2
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.2.0
python-memcached==1.57
PyYAML==3.13
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.1.16
texttable==0.9.1
traitlets==4.0.0
urllib3==1.22
websocket-client==0.48.0
wheel==0.24.0
